requirements.txt
httpx>=0.27.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.6.0
openai>=1.50.0
anthropic>=0.40.0
praw>=7.7.1
tenacity>=8.2.0
tzdata>=2024.1
